refactor(routes): share the TypeScript module-candidate resolver between the config extractor and the contract scan
Review: the re-export follower probed the emitted .js before its .ts/.tsx
source and skipped .cjs->.cts and index modules. Both static scans now use
one moduleCandidates() in routes/module-candidates.ts.
